sprint-3/step-2 fixed issues
