+++ src/pages/login.jsx
@@ -3,7 +3,7 @@ import { useHistory, Redirect, useLocation } from 'react-router-dom';
 import style from './login.module.css';
 import { Button } from '@ya.praktikum/react-developer-burger-ui-components';
 import { useDispatch, useSelector } from 'react-redux';
-import { loginUserThunk } from '../services/actions/actions';
+import { loginUserThunk } from '../services/actions/user';
 import { PasswordInput } from '../components/custom/input/password-input';
 import { EmailInput } from '../components/custom/input/email-input';
 import PropTypes from 'prop-types';
@@ -33,7 +33,6 @@ const LoginPage = () => {
     const login = useCallback(
         (e) => {
             e.preventDefault();
-            console.log('form ', form)
             dispatch(loginUserThunk(form));
         }, [form, dispatch]
     )
@@ -48,7 +47,7 @@ const LoginPage = () => {
         <div className={`${style.wrapper}`}>
             <h1 className={`text text_type_main-medium mb-6 ${style.title}`}>Вход</h1>
             {hasError && <span className={`mt-4 mb-4 text text_type_main-default ${style.error}`}>{error}</span>}
-            <form action="" className={`${style.form}`}>
+            <form action="" className={`${style.form}`} onSubmit={login}>
                 <div className="mb-6">
                     <EmailInput onChange={onChange} value={form.email} name={'email'} />
                 </div>
@@ -56,7 +55,7 @@ const LoginPage = () => {
                     <PasswordInput onChange={onChange} value={form.password} name={'password'} />
                 </div>
                 <div className="mb-20">
-                    <Button className={`${style.button}`} type="primary" size="medium" onClick={login}>Войти</Button>
+                    <Button className={`${style.button}`} type="primary" size="medium" >Войти</Button>
                 </div>
             </form>
             <div className={` ${style.text__wrapper} mb-4`}>
@@ -84,7 +83,7 @@ const LoginPage = () => {
 LoginPage.propTypes = {
     hasError: PropTypes.bool,
     error: PropTypes.string,
-    isAuthenticated: PropTypes.bool.isRequired
+    isAuthenticated: PropTypes.bool
 }
 
 export default LoginPage;
