+++ src/services/API.jsx
@@ -2,68 +2,73 @@ import { getCookie } from './utils';
 
 export const URL = 'https://norma.nomoreparties.space/api';
 
+function checkResponse(res) {
+    if (res.ok) {
+        return res.json();
+    }
+    return Promise.reject(`Ошибка ${res.status}`);
+}
+
 export async function fetchIngredients() {
-    const res = await fetch(URL + '/ingredients');
-    const data = await res.json();
+    const data = await fetch(URL + '/ingredients').then(checkResponse);
     return data
 }
 
 export async function fetchOrderIngredients(orderItems) {
     const body = { ingredients: orderItems };
-    const res = await fetch(URL + '/orders', {
+    const data = await fetch(URL + '/orders', {
         method: "POST",
         headers: {
             'Content-Type': 'application/json'
         },
         body: JSON.stringify(body)
-    });
-    const data = await res.json();
+    }).then(checkResponse);
+
     return data
 }
 
 export async function resetPasswordRequest(email) {
-    const res = await fetch(URL + '/password-reset', {
+    const data = await fetch(URL + '/password-reset', {
         method: "POST",
         headers: {
             'Content-Type': 'application/json'
         },
         body: JSON.stringify(email)
-    });
-    const data = await res.json();
+    }).then(checkResponse);
+
     return data
 }
 
 export async function saveNewPasswordRequest(form) {
     console.log(form)
-    const res = await fetch(URL + '/password-reset/reset', {
+    const data = await fetch(URL + '/password-reset/reset', {
         method: "POST",
         headers: {
             'Content-Type': 'application/json'
         },
         body: JSON.stringify(form)
-    });
-    console.log('res ', res)
-    const data = await res.json();
+    }).then(checkResponse);
+
     console.log('data response ', data)
     return data
 }
 
 export async function createUser(form) {
     console.log('form createUser ', form)
-    const res = await fetch(URL + '/auth/register', {
+    const data = await fetch(URL + '/auth/register', {
         method: "POST",
         headers: {
             'Content-Type': 'application/json'
         },
         body: JSON.stringify(form)
-    });
-    const data = await res.json();
+    }).then(checkResponse);
+
     console.log('createUser API data ', data)
     return data
 }
 
 export async function loginUser(form) {
-    const res = await fetch(URL + '/auth/login', {
+    const data = await fetch(URL + '/auth/login', {
         method: "POST",
         method: 'POST',
         mode: 'cors',
@@ -75,8 +80,8 @@ export async function loginUser(form) {
         redirect: 'follow',
         referrerPolicy: 'no-referrer',
         body: JSON.stringify(form)
-    });
-    const data = await res.json();
+    }).then(checkResponse)
+    console.log('login ', data)
     return data
 }
 
@@ -85,22 +90,20 @@ export async function logoutUser() {
     const body = {
         token: token
     }
-    const res = await fetch(URL + '/auth/logout', {
+    const data = await fetch(URL + '/auth/logout', {
         method: "POST",
         headers: {
             'Content-Type': 'application/json'
         },
         body: JSON.stringify(body)
-    });
+    }).then(checkResponse);
 
-    const data = await res.json();
-    console.log('log out data ', data)
     return data
 }
 
 
 export async function getUserInfo() {
-    const res = await fetch(URL + '/auth/user', {
+    const data = await fetch(URL + '/auth/user', {
         method: "GET",
         mode: 'cors',
         cache: 'no-cache',
@@ -111,11 +114,9 @@ export async function getUserInfo() {
         },
         redirect: 'follow',
         referrerPolicy: 'no-referrer'
-    });
+    }).then(checkResponse);
 
-
-    const data = await res.json();
-    console.log('fetch User ', data)
+    //console.log('fetch User ', data)
     return data;
 }
 
@@ -124,32 +125,28 @@ export async function refreshToken(token) {
         token: token
     }
 
-    const res = await fetch(URL + '/auth/token', {
+    const data = await fetch(URL + '/auth/token', {
         method: "POST",
         headers: {
             'Content-Type': 'application/json'
         },
         body: JSON.stringify(body)
-    });
-
-    const data = await res.json();
-    console.log('fetch refreshToken ', data)
+    }).then(checkResponse);
+    //console.log('fetch refreshToken ', data)
     return data
 }
 
 
 export async function updateUser(form) {
 
-    const res = await fetch(URL + '/auth/user', {
+    const data = await fetch(URL + '/auth/user', {
         method: "PATCH",
         headers: {
             'Content-Type': 'application/json',
             Authorization: 'Bearer ' + getCookie('accessToken')
         },
         body: JSON.stringify(form)
-    });
-
-    const data = await res.json();
+    }).then(checkResponse);
 
     return data
 }
